Various updates
Random strong password generation and encrypted storage of wallet opening and spending passwords has been implemented.
Added automatic login support for last used wallet information during active session.
Added address book.
If dApp and wallet are on different networks, notification message is displayed on dApp and wallet.
After entering the password on the login screen, it is possible to log in with the "Enter" key.
Added address verification (NAV/xNAV) feature in the sending section.
When creating an NFT collection, an error message is displayed if there is no xNAV in the account.